refactoring and tests completed
