+++ client/api/foods.test.js
@@ -1,4 +1,4 @@
-import { fetchFoodById, fetchFoods } from './foods'
+import { fetchFoodById } from './foods'
 
 jest.mock('./requestor', () => {
   // requestor exports a function, so we must do so here too
@@ -9,29 +9,7 @@ jest.mock('./requestor', () => {
       return Promise.reject(new Error('id not found'))
     } else {
       return Promise.resolve({
-        body: [
-          {
-            id: 1,
-            name: 'Lamb',
-            category_id: 5,
-            carbonOutput: 20.85,
-            waterUsage: 8763
-          },
-          {
-            id: 2,
-            name: 'Beef',
-            category_id: 5,
-            carbonOutput: 13.78,
-            waterUsage: 15000
-          },
-          {
-            id: 3,
-            name: 'Turkey',
-            category_id: 5,
-            carbonOutput: 5.83,
-            waterUsage: 5291
-          }
-        ]
+        id: id
       })
     }
   }
@@ -48,22 +26,6 @@ describe('API client for food', () => {
   it('fetchFoodsById returns an error when expected', () => {
     expect.assertions(1) // otherwise a valid id would still pass
     return fetchFoodById(33).catch(e => {
-      expect(e.message).toBe('Error accessing foods api.')
-    })
-  })
-})
-
-describe('API client for all foods', () => {
-  it('fetchFoods returns all foods', () => {
-    return fetchFoods()
-      .then(foods => {
-        expect(foods).toHaveLength(3)
-      })
-  })
-
-  it('fetchFoods returns an error when expected', () => {
-    // expect.assertions(1) // otherwise a valid id would still pass
-    return fetchFoods().catch(e => {
       expect(e.message).toBe('Error accessing foods api.')
     })
   })
